Below Changes Done:
# Removed Unncessary Code Files
# Splitted CustomerService Class and MessageService Class
# Updated existing unit test cases
# Added few more test cases
