+++ CandidateCodeTest/Services/Interfaces/ICustomerService.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace CandidateCodeTest.Services
+﻿namespace CandidateCodeTest.Services
 {
     public interface ICustomerService
     {
